pyproject.toml
[project]
name = "worldpop-stac-api"
version = "1.0.0"
description = "WorldPop STAC API implementation"
authors = [
    { name = "Borys Nosatiuk", email = "b.nosatiuk@soton.ac.uk" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.27.0",
    "stac-fastapi.types~=3.0.0",
    "stac-fastapi.api~=3.0.0",
    "stac-fastapi.extensions~=3.0.0",
    "stac-pydantic~=3.1.0",
    "motor>=3.4.0",
    "orjson>=3.9.10",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "attrs>=23.2.0",
    "pygeofilter>=0.2.0",
    "brotli-asgi>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "pre-commit>=3.6.0",
]

aws = [
    "mangum>=0.17.0",  # AWS Lambda support
]

[project.urls]
Homepage = "https://github.com/wpgp/worldpop-stac-api"
Documentation = "https://github.com/wpgp/worldpop-stac-api#readme"
Repository = "https://github.com/wpgp/worldpop-stac-api.git"
Issues = "https://github.com/wpgp/worldpop-stac-api/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "D",   # pydocstyle
    "B",   # flake8-bugbear
]
ignore = ["D203", "D212"]

[tool.ruff.per-file-ignores]
"tests/*" = ["D"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.isort]
known-first-party = ["wpstac"]
known-third-party = ["stac_fastapi", "stac_pydantic"]

[tool.mypy]
target-version = "py312"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "stac_fastapi.*",
    "stac_pydantic.*",
    "motor.*",
    "orjson.*",
    "pygeofilter.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=wpstac --cov-report=term-missing"
python_files = "test_*.py"